[BUG][ENH] Renaming and model changes
Due to the namespacing used, it was safer to modify the directory names
so that there are no underscores or dashes as this seemed to create
problems setting up the packages.
Further progress has been added to the model and model adaptor which now
returns the correct outputs for a post/get request. A test should be
created to check the quality of the outputs against a known standard.